+++ R/cut_period.R
@@ -33,9 +33,22 @@ cut_period <- function(data, start, end, len, unit = c("day", "week", "month", "
   stopifnot(is.data.frame(data))
 
   err <- data %>%
-    dplyr::filter({{ start }} > {{ end }}) %>%
-    nrow()
-  if (err > 0) stop("Some start date is later than end date.")
+    dplyr::filter(dplyr::if_any(c({{ start }}, {{ end }}), is.na))
+
+  if (nrow(err) > 0) {
+    cat("\nThe following records have missing dates:\n")
+    print(err)
+    stop("Some start or end dates are missing.")
+  }
+
+  err <- data %>%
+    dplyr::filter({{ start }} > {{ end }})
+
+  if (nrow(err) > 0) {
+    cat("\nThe following records have start > end:\n")
+    print(err)
+    stop("Some start > end.")
+  }
 
   unit <- rlang::arg_match0(unit, c("day", "week", "month", "quarter", "year"))
 
@@ -49,7 +62,7 @@ cut_period <- function(data, start, end, len, unit = c("day", "week", "month", "
   data %>%
     dplyr::mutate(
       segment_start = purrr::map2({{ start }}, {{ end }}, ~ seq(.x, .y, by = paste(len, unit))),
-      segment_end = purrr::map2(.data$segment_start, {{ end }}, ~ .x %>% dplyr::lead(, default = .y + lubridate::days(1)) - lubridate::days(1)),
+      segment_end = purrr::map2(.data$segment_start, {{ end }}, ~ .x %>% dplyr::lead(, default = .y %m+% lubridate::days(1)) %m-% lubridate::days(1)),
       segment_id = purrr::map(.data$segment_start, ~ seq_along(lengths(.x)))
     ) %>%
     tidyr::unnest(cols = dplyr::any_of(new_cols))