+++ R/restrict_dates_sql.R
@@ -10,8 +10,18 @@ restrict_dates.tbl_sql <- function(data, clnt_id, date_var, n, apart = NULL, wit
   clnt_id <- rlang::as_name(rlang::enquo(clnt_id))
   date_var <- rlang::as_name(rlang::enquo(date_var))
 
+  # check missing date_var
+  check_null <- data %>%
+    dplyr::count(no_date = is.na(.data[[date_var]])) %>%
+    dplyr::collect()
+  if (nrow(check_null) > 1) {
+    warning("Removed ", check_null[2, 2], " records with missing date_var")
+    data <- data %>%
+      dplyr::filter(!is.na(.data[[date_var]]))
+  }
+
   # place holder for temp var names
-  flag_restrict_date <- temp_nm_keep_max <- temp_nm_keep_cum <- temp_nm_diff <- temp_nm_gap <- temp_nm_drank <- temp_nm_drank_diff <- temp_nm_drank_gap <- NULL
+  flag_restrict_date <- temp_nm_keep_max <- temp_nm_keep_cum <- temp_nm_diff <- temp_nm_gap <- temp_nm_dup <- temp_nm_last_date <- NULL
 
   n <- as.integer(n)
 
@@ -22,12 +32,6 @@ restrict_dates.tbl_sql <- function(data, clnt_id, date_var, n, apart = NULL, wit
       # see if_dates for detail
       keep <- dplyr::collect(data) %>%
         restrict_dates.data.frame(clnt_id = !!clnt_id, date_var = !!date_var, n = n, apart = apart, within = within, flag_at = flag_at, mode = mode, dup.rm = dup.rm, ...)
-      # dplyr::group_by(.data[[clnt_id]]) %>%
-      # dplyr::arrange(.data[[clnt_id]], .data[[date_var]]) %>%
-      # dplyr::mutate(temp_nm_keep = if_dates(.data[[date_var]], n, apart, within, dup.rm, ...)) %>%
-      # dplyr::filter(temp_nm_keep) %>%
-      # dplyr::select(-dplyr::starts_with("temp_")) %>%
-      # dplyr::ungroup()
     }
   } else {
     has_uid <- !rlang::quo_is_null(rlang::enquo(uid))
@@ -41,31 +45,44 @@ restrict_dates.tbl_sql <- function(data, clnt_id, date_var, n, apart = NULL, wit
     stopifnot(is.wholenumber(within))
     within <- as.integer(within)
 
-    keep <- data %>%
-      dplyr::group_by(.data[[clnt_id]]) %>%
-      dbplyr::window_order(.data[[date_var]], .data[[uid]]) %>%
-      dplyr::mutate(
-        temp_nm_drank = dplyr::dense_rank(.data[[date_var]])
+    # to get consecutive unique dates, flag duplicate and use it to sort
+    if (dup.rm) {
+      data <- data %>%
+        dplyr::group_by(.data[[clnt_id]]) %>%
+        dbplyr::window_order(.data[[date_var]], .data[[uid]]) %>%
+        dplyr::mutate(
+          temp_nm_last_date = dplyr::lag(.data[[date_var]]),
+          temp_nm_dup = dplyr::case_when(
+            .data[[date_var]] == temp_nm_last_date ~ 1L,
+            is.na(temp_nm_last_date) ~ 0L,
+            .default = 0L
+          )
         )
-    #browser()
+      keep <- data %>%
+        dplyr::group_by(.data[[clnt_id]], temp_nm_dup) %>%
+        dbplyr::window_order(.data[[date_var]], .data[[uid]])
+    } else {
+      keep <- data %>%
+        dplyr::group_by(.data[[clnt_id]]) %>%
+        dbplyr::window_order(.data[[date_var]], .data[[uid]])
+    }
+    # browser()
 
-    switch (flag_at,
+    switch(flag_at,
       left = {
         keep <- keep %>%
           dplyr::mutate(
-            temp_nm_diff = dplyr::lead(.data[[date_var]], n = local(n - 1L)),
-            temp_nm_drank_diff = dplyr::lead(temp_nm_drank, n = local(n - 1L))
+            temp_nm_diff = dplyr::lead(.data[[date_var]], n = local(n - 1L))
           )
       },
       right = {
         keep <- keep %>%
           dplyr::mutate(
-            temp_nm_diff = dplyr::lag(.data[[date_var]], n = local(n - 1L)),
-            temp_nm_drank_diff = dplyr::lag(temp_nm_drank, n = local(n - 1L))
+            temp_nm_diff = dplyr::lag(.data[[date_var]], n = local(n - 1L))
           )
       }
     )
-    #browser()
+    # browser()
 
     # SQL server does not accept subtracting dates
     is_mssql_mysql <- stringr::str_detect(dbplyr::remote_con(data) %>% class(), "SQL Server|Maria") %>% any()
@@ -81,47 +98,36 @@ restrict_dates.tbl_sql <- function(data, clnt_id, date_var, n, apart = NULL, wit
           temp_nm_gap = abs(temp_nm_diff - .data[[date_var]])
         )
     }
-    #browser()
+    # browser()
 
-    keep <- keep %>%
-      dplyr::mutate(
-        # the translation for any() failed on SQL server again
-        # temp_nm_keep = any(temp_nm_gap <= within, na.rm = TRUE)
-        temp_nm_drank_gap = abs(temp_nm_drank_diff - temp_nm_drank),
-        flag_restrict_date = dplyr::case_when(
-          temp_nm_gap == 0L ~ 0L,
-          temp_nm_gap <= within & temp_nm_drank_gap == n - 1L ~ 1L,
-          is.na(temp_nm_gap) ~ 0L,
-          .default = 0L
+    if (dup.rm) {
+      keep <- keep %>%
+        dplyr::mutate(
+          # the translation for any() failed on SQL server again
+          # temp_nm_keep = any(temp_nm_gap <= within, na.rm = TRUE)
+          flag_restrict_date = dplyr::case_when(
+            temp_nm_dup == 1L ~ NA,
+            temp_nm_gap <= within ~ 1L,
+            is.na(temp_nm_gap) ~ 0L,
+            .default = 0L
+          )
         )
-      )
+      # sort back to original order
+      keep <- keep %>%
+        dplyr::group_by(.data[[clnt_id]]) %>%
+        dbplyr::window_order(.data[[date_var]], .data[[uid]]) %>%
+        tidyr::fill(flag_restrict_date, .direction = "down")
 
-    # if (strict_start) {
-    #   keep <- keep %>%
-    #     dplyr::mutate(temp_nm_keep_cum = cummax(temp_nm_keep))
-    #   switch(mode,
-    #     "flag" = {
-    #       keep <- keep %>%
-    #         dplyr::mutate(flag_restrict_date = temp_nm_keep_cum)
-    #     },
-    #     "filter" = {
-    #       keep <- keep %>% dplyr::filter(temp_nm_keep_cum > 0L)
-    #     }
-    #   )
-    # } else {
-    #   keep <- keep %>%
-    #     dplyr::mutate(temp_nm_keep_max = max(temp_nm_keep, na.rm = TRUE))
-    #   switch(mode,
-    #     "flag" = {
-    #       keep <- keep %>%
-    #         dplyr::mutate(flag_restrict_date = temp_nm_keep_max)
-    #     },
-    #     "filter" = {
-    #       keep <- keep %>% dplyr::filter(temp_nm_keep_max >= 1L)
-    #     }
-    #   )
-    # }
-    #browser()
+    } else {
+      keep <- keep %>%
+        dplyr::mutate(
+          flag_restrict_date = dplyr::case_when(
+            temp_nm_gap <= within ~ 1L,
+            is.na(temp_nm_gap) ~ 0L,
+            .default = 0L
+          )
+        )
+    }
 
     if (mode == "filter") {
       keep <- keep %>% dplyr::filter(max(flag_restrict_date, na.rm = TRUE) > 0L)