+++ tests/testthat/test-cut_period.R
@@ -5,7 +5,7 @@ test_that("basic use works", {
   df <- df %>%
     dplyr::mutate(
       add = ans_rows,
-      end_date = purrr::map2_vec(dates, add, ~ .x + lubridate::period(.y, units = unt) - lubridate::days(1)))
+      end_date = purrr::map2_vec(dates, add, ~ .x %m+% lubridate::period(.y, units = unt) %m-% lubridate::days(1)))
   out_df <- df %>% cut_period(start = dates, end = end_date, len = 1, unit = unt)
   expect_equal(nrow(out_df), sum(ans_rows))
   expect_in(names(df), names(out_df))
@@ -18,7 +18,7 @@ test_that("date transform works", {
   df <- df %>%
     dplyr::mutate(
       add = ans_rows,
-      end_date = purrr::map2_vec(dates, add, ~ .x + lubridate::period(.y, units = unt) - lubridate::days(1)),
+      end_date = purrr::map2_vec(dates, add, ~ .x %m+% lubridate::period(.y, units = unt) %m-% lubridate::days(1)),
       dplyr::across(dplyr::where(lubridate::is.Date), as.character))
   expect_error(cut_period(df, start = dates, end = end_date, len = 1, unit = unt), "is not Date")
   out_df <- df %>% cut_period(start = dates, end = end_date, len = 1, unit = unt, .dt_trans = as.Date)
@@ -31,5 +31,14 @@ test_that("edge case start > end", {
   df <- df %>%
     dplyr::mutate(
       end_date = dates - 7)
-  expect_error(df %>% cut_period(start = dates, end = end_date, len = 1, unit = unt), "later")
+  expect_error(df %>% cut_period(start = dates, end = end_date, len = 1, unit = unt), ">")
+})
+
+test_that("edge case missing start", {
+  df <- letters_n()
+  unt <- "month"
+  df <- df %>%
+    dplyr::mutate(
+      end_date = ifelse(uid != 1, dates + 7, NA))
+  expect_error(df %>% cut_period(start = dates, end = end_date, len = 1, unit = unt), "missing")
 })