+++ tests/testthat/test-if_dates.R
@@ -7,8 +7,8 @@ test_that("basic case works", {
 
 test_that("dup.rm works", {
   dates <- c(rep(as.Date("2023-12-01"), 3), as.Date("2023-12-08"), as.Date("2023-12-16"))
-  expect_true(if_dates(dates, 5, 0, 30, dup.rm = FALSE))
-  expect_false(if_dates(dates, 5, 0, 30, dup.rm = TRUE))
+  expect_true(if_dates(dates, 5, within = 30, dup.rm = FALSE))
+  expect_false(if_dates(dates, 5, within = 30, dup.rm = TRUE))
 })
 
 test_that("input checks work", {
@@ -49,6 +49,24 @@ test_that("detail works", {
   expect_equal(out, ans)
 })
 
+test_that("detail and dup.rm edge case", {
+  x <- as.Date(c("2010-01-01", "2010-01-01", "2012-05-03", "2012-05-03", "2015-01-07", "2015-02-01", "2017-02-08", "2017-05-07"))
+  ans_n2_dupT <- c(FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, TRUE, FALSE)
+  w <- 365
+  out <- if_dates(x, n = 2, within = w, detail = TRUE)
+  expect_equal(out, ans_n2_dupT)
+  ans_n2_dupF <- c(TRUE, FALSE, TRUE, FALSE, TRUE, FALSE, TRUE, FALSE)
+  out <- if_dates(x, n = 2, within = w, detail = TRUE, dup.rm = FALSE)
+  expect_equal(out, ans_n2_dupF)
+  x <- as.Date(c("2010-01-01", "2010-01-01", "2010-05-03", "2010-05-03", "2015-01-07", "2017-02-01", "2017-02-08", "2017-05-07"))
+  ans_n3_dupT <- c(FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE)
+  out <- if_dates(x, n = 3, within = w, detail = TRUE, dup.rm = TRUE)
+  expect_equal(out, ans_n3_dupT)
+  ans_n3_dupF <- c(TRUE, TRUE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE)
+  out <- if_dates(x, n = 3, within = w, detail = TRUE, dup.rm = FALSE)
+  expect_equal(out, ans_n3_dupF)
+})
+
 test_that("align works", {
   x <- as.Date(c("2010-01-01", "2012-05-03", "2015-01-07", "2015-02-01", "2017-02-08", "2017-05-07"))
   ans <- c(FALSE, FALSE, FALSE, TRUE, FALSE, TRUE)
@@ -63,5 +81,13 @@ test_that("align with apart works", {
   w <- 365
   a <- 60
   out <- if_dates(x, n = 2, apart = a, within = w, detail = TRUE, align = "right")
+  expect_equal(out, ans)
+})
+
+test_that("sort back works", {
+  x <- as.Date(c("2017-05-07", "2015-02-01", "2010-01-01", "2012-05-03", "2015-01-07",  "2017-02-08"))
+  ans <- c(TRUE, TRUE, FALSE, FALSE, FALSE, FALSE)
+  w <- 365
+  out <- if_dates(x, n = 2, within = w, detail = TRUE, align = "right")
   expect_equal(out, ans)
 })