+++ tests/testthat/test-restrict_dates_dataframe.R
@@ -36,3 +36,11 @@ test_that("flag_at works", {
   expect_equal(output_df$dates, x)
   expect_setequal(output_df$dates[output_df$flag_restrict_date == 1], x[ans_right])
 })
+
+test_that("check missing dates works", {
+  n <- sample(2:5, 1)
+  within <- sample(30:365, 1)
+  df <- make_test_dat()
+  df$dates[sample(1:nrow(df), 5)] <- NA
+  expect_warning(restrict_dates(df, clnt_id, dates, n, within = within, uid = uid), "Removed 5 records")
+})