+++ tests/testthat/test-restrict_dates_sql.R
@@ -18,10 +18,11 @@ test_that("within only works", {
   output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid, mode = "filter") %>% dplyr::collect()
   ans_id <- test_apart_within(df, n, within = within)
   expect_setequal(output_df$clnt_id, ans_id)
-  #also test uid error
+  # also test uid error
   expect_error(restrict_dates(db, clnt_id, dates, n, within = within, mode = "filter"), "uid")
   # also test mode
-  output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid) %>% dplyr::collect() %>%
+  output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid) %>%
+    dplyr::collect() %>%
     dplyr::filter(flag_restrict_date > 0)
   expect_setequal(output_df$clnt_id, ans_id)
 })
@@ -29,7 +30,7 @@ test_that("within only works", {
 test_that("do not count same date works for database", {
   n <- 3
   within <- 365
-  df <- data.frame(uid = n*2+1*3, clnt_id = rep(1:3, each = n*2+1), dates = rep(c(rep(as.Date("2020-01-01"),n*2-1), as.Date("2020-01-01")+within+1, as.Date("2020-01-01")+within*2+1), 3))
+  df <- data.frame(uid = n * 2 + 1 * 3, clnt_id = rep(1:3, each = n * 2 + 1), dates = rep(c(rep(as.Date("2020-01-01"), n * 2 - 1), as.Date("2020-01-01") + within + 1, as.Date("2020-01-01") + within * 2 + 1), 3))
   db <- dbplyr::memdb_frame(df)
   output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid, mode = "filter") %>% dplyr::collect()
   ans_id <- test_apart_within(df, n, within = within)
@@ -47,4 +48,35 @@ test_that("flag_at works", {
   output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid, flag_at = "right", mode = "filter") %>% dplyr::collect()
   ans_right <- c(FALSE, FALSE, FALSE, TRUE, FALSE, TRUE)
   expect_setequal(output_df$dates[output_df$flag_restrict_date == 1] %>% as.Date(origin = "1970-01-01"), x[ans_right])
+})
+
+test_that("count same date works for database", {
+  n <- 3
+  within <- 365
+  df <- data.frame(uid = n * 2 + 1 * 3, clnt_id = rep(1:3, each = n * 2 + 1), dates = rep(c(rep(as.Date("2020-01-01"), n * 2 - 1), as.Date("2020-01-01") + within + 1, as.Date("2020-01-01") + within * 2 + 1), 3))
+  db <- dbplyr::memdb_frame(df)
+  output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid, mode = "filter", dup.rm = FALSE) %>% dplyr::collect()
+  ans <- df %>%
+    dplyr::mutate(ans = if_date(dates, n = n, within = within, dup.rm = FALSE)) %>%
+    dplyr::filter(ans)
+  expect_setequal(output_df$clnt_id, ans$clnt_id)
+})
+
+test_that("check missing dates works", {
+  n <- sample(2:5, 1)
+  within <- sample(30:365, 1)
+  df <- make_test_dat()
+  df$dates[sample(1:nrow(df), 5)] <- NA
+  db <- memdb_tbl(df)
+  expect_warning(restrict_dates(db, clnt_id, dates, n, within = within, uid = uid), "Removed 5 records")
+})
+
+test_that("edge case duplicated dates and n > 2", {
+  x <- as.Date(c("2010-01-01", "2010-01-01", "2010-03-01", "2010-03-01", "2010-04-08"))
+  within <- 365
+  n <- 3
+  ans <- if_date(x, n, within = within, detail = TRUE)
+  db <- dbplyr::memdb_frame(clnt_id = 1, dates = x, uid = 1:length(x))
+  output_df <- restrict_dates(db, clnt_id, dates, n, within = within, uid = uid, flag_at = "right") %>% dplyr::collect()
+  expect_setequal(output_df$flag_restrict_date, as.numeric(ans))
 })